Change the QoS for the gre tunnel
Download doesn't change. The upload part is splitted in 2: the upload
which are done directly through the home IP (torrents and the connection
with the gre tunnel), and the main traffic which go out through the gre
tunnel. The GRE Tunnel has to be prioritary, as almost everything is leaving
through it.
Then a QoS is applicated in the tunnel.